% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bach_chorales_1.R
\docType{data}
\name{bach_chorales_1b}
\alias{bach_chorales_1b}
\title{Bach chorales (1b)}
\description{
This dataset comprises 370 chorales by J. S. Bach.
}
\details{
This dataset was constructed from data downloaded from
\href{http://kern.humdrum.org/cgi-bin/browse?l=/371chorales}{KernScores}.
The original KernScores data comprises full polyphonic scores.
These scores were converted to chord sequences
using the algorithm of
\insertCite{Pardo2002;textual}{hcorp},
with an expanded chord dictionary,
and colocating segment boundaries with metrical beat locations
as estimated from time signatures.

According to KernScores, these chorales were collected
after J. S. Bach's death by his son C. P. E. Bach,
and finished by Kirnberger, a student of J. S. Bach,
after C. P. E. Bach's death.
The KernScores dataset was constructed from
the fourth edition of the chorales by Breitkopf & Hartel, c. 1875,
entitled "371 four-part chorale chants of Johann Sebastian Bach",
retypeset c. 1915 as Edition Breitkopf 10,
and reprinted by Associated Music Publishers, Inc., New York in c. 1940.

The dataset uses S3 classes from the \code{hrep} package,
specifically:
\code{\link[hrep]{pc_chord}},
for representing chords;
\code{\link[hrep]{vec}},
for representing chord sequences;
\code{\link[hrep]{corpus}},
for representing the corpus.

Each chord is represented as a 'pitch-class chord',
corresponding to a bass pitch class and a set of non-bass pitch classes.
}
\note{
The dataset excludes number 150, the only chorale not in four parts.
}
\references{
\insertAllCited{}
}
\keyword{data}
