% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_1.R
\docType{data}
\name{classical_1b}
\alias{classical_1b}
\title{Classical corpus 1b}
\description{
Equivalent to \code{\link{classical_1}} but without applying
any chord labelling algorithm
(i.e. instead applying full expansion, a.k.a. salami slicing).
The dataset has a 'pitch chord' representation (\code{\link[hrep]{pi_chord}});
a pitch chord corresponds to the set of MIDI note numbers in the sonority.
}
\keyword{data}
