% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus.R
\name{get_corpus}
\alias{get_corpus}
\title{Get corpus}
\usage{
get_corpus(name)
}
\arguments{
\item{name}{(Character scalar) Name of the corpus to access.
Available names are listed in \code{corpora}.}
}
\description{
This function provides an alternative to accessing corpora
like \code{hcorp::classical_1}.
Instead, the same corpus can be accessed by calling
\code{get_corpus("classical")}.
}
