% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular_1.R
\docType{data}
\name{popular_1}
\alias{popular_1}
\title{Popular corpus 1}
\description{
This dataset provides chord sequences for 739 pieces from the
McGill Billboard corpus \insertCite{Burgoyne2011}{hcorp}.
These pieces were sampled from the Billboard magazine's
United States "Hot 100" chart between 1958 and 1991,
and transcribed by expert musicians.
}
\details{
While the original dataset comprises notated chord symbols,
this version of the dataset provides a more explicit representation,
where each chord is represented as a pitch class set and a labelled
bass pitch class.
This dataset was used in
\insertCite{Harrison2018;textual}{hcorp}.

The dataset uses S3 classes from the \code{hrep} package,
specifically:
\code{\link[hrep]{pc_chord}},
for representing chord symbols;
\code{\link[hrep]{coded_vec}},
for coding chord sequences as integer vectors;
\code{\link[hrep]{corpus}},
for representing the corpus.

Each chord is represented as a 'pitch-class chord':
this is defined as a pitch class set with a known bass pitch class.
}
\note{
\insertCite{Harrison2018;textual}{hcorp}
discarded inversion information when modelling this dataset.
}
\references{
\insertAllCited{}
}
\keyword{data}
